kudo app: refactor denomination
