+++ native/examples/kudo_application/kudo_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kudo_core"
+name = "kudo-core"
 version = "0.1.0"
 edition = "2021"
