+++ native/examples/kudo_application/kudo_core/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod denomination_logic_witness;
+pub mod denomination;
 pub mod kudo_logic_witness;
 pub mod receive_logic_witness;
 pub mod utils;
