+++ native/examples/kudo_application/kudo_examples/simple_denomination/core/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "denomination-core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+aarm_core = { path = "../../../../../aarm_core" }
+kudo-core = { path = "../../../kudo_core" }
+serde = { version = "1.0.197", default-features = false }
