+++ native/examples/kudo_application/kudo_examples/simple_denomination/methods/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "denomination_logic"
+name = "denomination-logic-circuit"
 version = "0.1.0"
 edition = "2021"
