+++ native/examples/kudo_application/kudo_examples/simple_denomination/methods/src/lib.rs
