+++ native/examples/kudo_application/kudo_tx/Cargo.toml
@@ -1,16 +1,15 @@
 [package]
-name = "kudo_tx"
+name = "kudo-tx"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
 aarm_core = { path = "../../../aarm_core" }
-kudo_core = { path = "../kudo_core" }
+kudo-core = { path = "../kudo_core" }
 aarm = { path = "../../../aarm" }
 serde = { version = "1.0.197", default-features = false }
 risc0-zkvm = "1.2.5"
 compliance_circuit = { path = "../../compliance_circuit/methods", package = 'methods' }
 trivial_logic = { path = "../../trivial_logic/trivial_logic" }
 kudo_logic = { path = "../kudo_logic" }
-denomination_logic = { path = "../denomination_logic" }
 receive_logic = { path = "../receive_logic" }
