+++ native/examples/kudo_application/receive_logic/guest/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 risc0-zkvm = { version = "1.2.5", default-features = false, features = ['std'] }
-kudo_core = { path = "../../kudo_core" }
+kudo-core = { path = "../../kudo_core" }
 
 [patch.crates-io]
 # Placing this patch statement in the workspace Cargo.toml will add RISC Zero SHA-256 accelerator
