+++ native/examples/kudo_application/rust-toolchain.toml
@@ -1,4 +1,5 @@
 [toolchain]
-channel = "stable"
-components = ["rustfmt", "rust-src"]
+channel = "1.85"
+components = ["clippy", "rustfmt", "rust-src"]
+targets = ["wasm32-unknown-unknown"]
 profile = "minimal"
